refactor(demo): navigation
Top menu entries of navigation tree do not have to be a XHTML. Instead the menu entry can be set with a menuEntry.properties file which contains a "label" and "icon" value.
The tree is now completely expanded. In the future, sub-sub elements should be rendered as a second menu on the site.
Replace useless loplevel XHTMLs with menuEntry.properties.